+++ crates/module-system/sov-modules-api/src/transaction/mod.rs
@@ -33,6 +33,7 @@ use crate::{
     MeteredSignature, Spec,
 };
 
+pub mod legacy_v0;
 mod signing_payload;
 #[cfg(test)]
 mod tests;